feat(auth): 更新账户状态和密码强度字段类型为Integer
- 将AuthsAccount实体中的status字段从Short改为Integer
- 将AuthsAccount实体中的passwordStrength字段从Short改为Integer
- 删除系统字典表相关文件（实体、控制器、DTO、Mapper及其实现）
- 优化菜单模块，增加根据账户ID获取菜单功能
- 新增接入层控制器和服务实现，支持验证码、登录、注册、登出及密码重置功能
- 菜单实体新增翻译注解和子级字段，支持树形结构展示
- 增加基于账户ID查询菜单的SQL映射和业务逻辑实现

+++ dream-app/app-system/src/main/java/io/jiangbyte/app/base/auths/account/entity/AuthsAccount.java
@@ -49,10 +49,10 @@ public class AuthsAccount extends BaseEntity {
     private String telephone;
 
     @Schema(description = "账户状态：0-正常, 1-锁定, 2-禁用")
-    private Short status;
+    private Integer status;
 
     @Schema(description = "密码强度等级：0-3")
-    private Short passwordStrength;
+    private Integer passwordStrength;
 
     @Schema(description = "最后修改密码的时间")
     private Date lastPasswordChange;
